add execute-in-vm and cleanup-vm task

+++ scripts/deploy-tasks.sh
@@ -17,26 +17,36 @@ DEPLOY_NAMESPACE="${DEPLOY_NAMESPACE:-$(oc project --short)}"
 oc project "${DEPLOY_NAMESPACE}" || true
 
 visit "${REPO_DIR}/tasks"
-  for TASK_NAME in *; do
-    CONFIG_FILE="${REPO_DIR}/configs/${TASK_NAME}.yaml"
-    MAIN_IMAGE="$(sed -n  's/^main_image *: *//p' "${CONFIG_FILE}")"
-    CUSTOM_IMAGE="${TASK_NAME_TO_IMAGE[${TASK_NAME}]}"
-
-    visit "${TASK_NAME}"
-      oc delete -f manifests 2> /dev/null || true
-      if [[ $SCOPE == "cluster" ]]; then
-        sed "s/TARGET_NAMESPACE/${DEPLOY_NAMESPACE}/" "manifests/${TASK_NAME}-cluster-rbac.yaml" | oc apply -f -
-      else
-        oc apply -f "manifests/${TASK_NAME}-namespace-rbac.yaml"
+  if [[ $# -eq 0 ]]; then
+    TASK_NAMES=(*)
+  else
+    TASK_NAMES=("$@")
+  fi
+  for TASK_NAME in ${TASK_NAMES[*]}; do
+    if echo "${TASK_NAME}" | grep -vE "^(${EXCLUDED_NON_IMAGE_MODULES})$"; then
+      if [ ! -d  "${TASK_NAME}" ]; then
+        continue
       fi
+      CONFIG_FILE="${REPO_DIR}/configs/${TASK_NAME}.yaml"
+      MAIN_IMAGE="$(sed -n  's/^main_image *: *//p' "${CONFIG_FILE}")"
+      CUSTOM_IMAGE="${TASK_NAME_TO_IMAGE[${TASK_NAME}]}"
 
-      for SUBTASK_NAME in $(ls manifests | grep -v rbac); do
-        if [[ -z ${CUSTOM_IMAGE} ]]; then
-          oc apply -f "manifests/${SUBTASK_NAME}"
+      visit "${TASK_NAME}"
+        oc delete -f manifests 2> /dev/null || true
+        if [[ $SCOPE == "cluster" ]]; then
+          sed "s/TARGET_NAMESPACE/${DEPLOY_NAMESPACE}/" "manifests/${TASK_NAME}-cluster-rbac.yaml" | oc apply -f -
         else
-          sed "s!${MAIN_IMAGE}!${CUSTOM_IMAGE}!g" "manifests/${SUBTASK_NAME}" | oc apply -f -
+          oc apply -f "manifests/${TASK_NAME}-namespace-rbac.yaml"
         fi
-      done
-    leave
+
+        for SUBTASK_NAME in $(ls manifests | grep -v rbac); do
+          if [[ -z ${CUSTOM_IMAGE} ]]; then
+            oc apply -f "manifests/${SUBTASK_NAME}"
+          else
+            sed "s!${MAIN_IMAGE}!${CUSTOM_IMAGE}!g" "manifests/${SUBTASK_NAME}" | oc apply -f -
+          fi
+        done
+      leave
+    fi
   done
 leave
